refactor: SortByDropdown 컴포넌트 공용으로 적용
